Modification post, commentaire OK__OK reglage v-show modification

+++ backend/controllers/post.js
@@ -51,17 +51,19 @@ exports.modifyPost = async (req, res, next) => {
 
         if (post.user_id === user_id) {
             if (req.file) {
-                const ancienneImage = post.imgUrl.split(/images/)[1];
-                fs.unlink(`./images/${ancienneImage}`, (err => {
-                    if (err) console.log(err);
-                    else {
-                        console.log("Ancienne image effacé " + ancienneImage);
-                    }
-                }))
-
-                let imgUrl = `${req.protocol}://${req.get('host')}/images/${req.file.filename}`;
-                await Post.update({ user_id, message: req.body.message, imgUrl }, { where: { id: postId } });
-                return res.status(200).json({ message: 'Message modifié avec succès !!' })
+                if (post.imgurl != null) {
+                    const ancienneImage = post.imgUrl.split(/images/)[1];
+                    fs.unlink(`./images/${ancienneImage}`, (err => {
+                        if (err) console.log(err);
+                        else {
+                            console.log("Ancienne image effacé " + ancienneImage);
+                        }
+                    }))
+                } else {
+                    let imgUrl = `${req.protocol}://${req.get('host')}/images/${req.file.filename}`;
+                    await Post.update({ user_id, message: req.body.message, imgUrl }, { where: { id: postId } });
+                    return res.status(200).json({ message: 'Message modifié avec succès !!' })
+                }
             }
             if (!req.file) {
                 console.log('No req.file');
